split data loading from infosheet
to different child components
